chore: seperate business logic and database operation
