+++ src/modules/article/article.service.ts
@@ -1,41 +1,33 @@
 import { Injectable } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
-import { Article } from './entities/article.entity';
-import { Repository } from 'typeorm';
 import { maxPageValidation, setPage } from '@common/utils/maxPageValidation';
 import { MetaData } from '@common/dto/metaData.dto';
 import { ArticleDto, ArticleListDto } from './dto/article.dto';
 import { utcToAsiaJakarta } from '@common/utils/timezone-converter';
 import { GraphQLError } from 'graphql';
+import { ArticleRepository } from './article.repository';
 
 @Injectable()
 export class ArticleService {
-  constructor(
-    @InjectRepository(Article)
-    private readonly articleRepository: Repository<Article>,
-  ) {}
+  constructor(private readonly articleRepository: ArticleRepository) {}
 
   async findAll(page: number, pageSize: number): Promise<ArticleListDto> {
     maxPageValidation(pageSize);
     const offset: number = setPage(page, pageSize);
 
     try {
-      const data = await this.articleRepository
-        .createQueryBuilder('articles')
-        .where('deleted_at IS NULL')
-        .orderBy('articles.id', 'DESC')
-        .skip(offset)
-        .take(pageSize)
-        .getManyAndCount();
+      const [data, total] = await this.articleRepository.findAll(
+        offset,
+        pageSize,
+      );
 
       const metaData: MetaData = {
         pageSize: pageSize,
         currentPage: page,
-        total: data[1],
-        totalPage: Math.ceil(data[1] / pageSize),
+        total,
+        totalPage: Math.ceil(total / pageSize),
       };
 
-      const articles: ArticleDto[] = data[0].map((item) => {
+      const articles: ArticleDto[] = data.map((item) => {
         const createdAt: string = utcToAsiaJakarta(item.createdAt);
         const updatedAt: string = utcToAsiaJakarta(item.updatedAt);
 
@@ -61,11 +53,7 @@ export class ArticleService {
 
   async findOne(id: number): Promise<ArticleDto> {
     try {
-      const article = await this.articleRepository
-        .createQueryBuilder('articles')
-        .where('articles.id = :id', { id })
-        .andWhere('articles.deleted_at IS NULL')
-        .getOne();
+      const article = await this.articleRepository.findById(id);
 
       if (!article) {
         throw new GraphQLError('Article not found.');
