+++ src/modules/user/user.module.ts
@@ -5,10 +5,11 @@ import { DateScalar } from 'src/libs/common/src/scalars/date.scalar';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from './entities/user.entity';
 import { AuthModule } from '@module/auth/auth.module';
+import { UserRepository } from './user.repository';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User]), AuthModule],
-  providers: [UserResolver, UserService, DateScalar],
+  providers: [UserResolver, UserService, DateScalar, UserRepository],
   exports: [UserService],
 })
 export class UserModule {}
